address review comments:
- move error to libs/errors
